requirements.txt
aiodataloader==0.1.2
aiofiles==0.3.2
asyncpg==0.15.0
atomicwrites==1.1.5
attrs==17.4.0
graphql-core==2.0
graphql-server-core==1.0.dev20170322001
httptools==0.0.11
more-itertools==4.2.0
pluggy==0.6.0
promise==2.1
py==1.5.3
pytest==3.6.0
pytest-asyncio==0.8.0
pytest-runner==4.2
Rx==1.6.1
sanic==0.7.0
Sanic-GraphQL==1.1.0
six==1.11.0
typing==3.6.4
ujson==1.35
uvloop==0.9.1
websockets==4.0.1
